refactor(predictions): Add async thunk for fetching static market data
