+++ src/views/Predictions/hooks/usePollPredictions.ts
@@ -2,12 +2,12 @@ import { useEffect, useRef } from 'react'
 import { useWeb3React } from '@web3-react/core'
 import { useAppDispatch } from 'state'
 import { useGetCurrentEpoch, useGetPredictionsStatus } from 'state/hooks'
-import { fetchRounds } from 'state/predictions'
+import { fetchMarketData, fetchRounds } from 'state/predictions'
 import { PredictionStatus } from 'state/types'
 
 const POLL_TIME_IN_SECONDS = 10
 
-const usePollRounds = () => {
+const usePollPredictions = () => {
   const timer = useRef<NodeJS.Timeout>(null)
   const dispatch = useAppDispatch()
   const { account } = useWeb3React()
@@ -24,6 +24,7 @@ const usePollRounds = () => {
       timer.current = setInterval(async () => {
         const liveAndCurrent = [currentEpoch, currentEpoch - 1]
         dispatch(fetchRounds(liveAndCurrent))
+        dispatch(fetchMarketData())
       }, POLL_TIME_IN_SECONDS * 1000)
     }
 
@@ -35,4 +36,4 @@ const usePollRounds = () => {
   }, [timer, account, status, currentEpoch, dispatch])
 }
 
-export default usePollRounds
+export default usePollPredictions
